feat(post): implement comment service: create and get root comments by postId APIs

+++ post_service/src/main/java/com/shibana/post_service/model/dto/response/PageResponse.java
@@ -11,7 +11,7 @@ public record PageResponse<T>(
         int page,
         int size,
         Boolean hasNext,
-        Long totalElements,
+        int totalElements,
         Integer totalPages,
         List<T> payload
 ) {
